This Is Coz A new change
This Is Coz A new changeThis Is Coz A new changeThis Is Coz A new change

+++ userbot/plugins/purge.py
@@ -12,6 +12,12 @@ from telethon.errors import rpcbaseerrors
 from userbot import BOTLOG, BOTLOG_CHATID, CMD_HELP
 from userbot.utils import register, errors_handler
 
+from userbot.uniborgConfig import Config
+
+
+BOTLOG = True
+BOTLOG_CHATID = Config.PRIVATE_GROUP_BOT_API_ID
+
 
 @register(outgoing=True, pattern="^.purge$")
 @errors_handler
@@ -19,16 +25,20 @@ async def fastpurger(purg):
     """ For .purge command, purge all messages starting from the reply. """
     chat = await purg.get_input_chat()
     msgs = []
+    itermsg = purg.client.iter_messages(chat, min_id=purg.reply_to_msg_id)
     count = 0
 
-    async for msg in purg.client.iter_messages(chat,
-                                               min_id=purg.reply_to_msg_id):
-        msgs.append(msg)
-        count = count + 1
-        msgs.append(purg.reply_to_msg_id)
-        if len(msgs) == 100:
-            await purg.client.delete_messages(chat, msgs)
-            msgs = []
+    if purg.reply_to_msg_id is not None:
+        async for msg in itermsg:
+            msgs.append(msg)
+            count = count + 1
+            msgs.append(purg.reply_to_msg_id)
+            if len(msgs) == 100:
+                await purg.client.delete_messages(chat, msgs)
+                msgs = []
+    else:
+        await purg.edit("`No message specified.`", )
+        return
 
     if msgs:
         await purg.client.delete_messages(chat, msgs)
@@ -111,46 +121,18 @@ async def editer(edit):
                                        "Edit query was executed successfully")
 
 
-@register(outgoing=True, pattern="^.sd")
-@errors_handler
-async def selfdestruct(destroy):
-    """ For .sd command, make seflf-destructable messages. """
-    message = destroy.text
-    counter = int(message[4:6])
-    text = str(destroy.text[6:])
-    await destroy.delete()
-    smsg = await destroy.client.send_message(destroy.chat_id, text)
-    await sleep(counter)
-    await smsg.delete()
-    if BOTLOG:
-        await destroy.client.send_message(BOTLOG_CHATID,
-                                          "sd query done successfully")
+
 
 
 CMD_HELP.update({
     'purge':
-    '.purge\
-        \nUsage: Purges all messages starting from the reply.'
+    ".purge\
+    \nUsage: Purges all messages starting from the reply.\
+    \n\n.purgeme <x>\
+    \nUsage: Deletes x amount of your latest messages.\
+    \n\n.del\
+    \nUsage: Deletes the message you replied to.\
+    \n\n.edit <newmessage>\
+    \nUsage: Replace your last message with <newmessage>."
 })
 
-CMD_HELP.update({
-    'purgeme':
-    '.purgeme <x>\
-        \nUsage: Deletes x amount of your latest messages.'
-})
-
-CMD_HELP.update({"del": ".del\
-\nUsage: Deletes the message you replied to."})
-
-CMD_HELP.update({
-    'edit':
-    ".edit <newmessage>\
-\nUsage: Replace your last message with <newmessage>."
-})
-
-CMD_HELP.update({
-    'sd':
-    '.sd <x> <message>\
-\nUsage: Creates a message that selfdestructs in x seconds.\
-\nKeep the seconds under 100 since it puts your bot to sleep.'
-})
